requirements.txt
Pillow==8.2.0
tqdm==4.60.0
